Extracted common functions into new file
