+++ pkg/runs/runsReset.go
@@ -114,50 +114,3 @@ func resetRun(runName string,
 
 	return err
 }
-
-func getRunIdFromRunName(runName string,
-	timeService utils.TimeService,
-	apiClient *galasaapi.APIClient,
-) (string, error) {
-	var err error
-	var runs []galasaapi.Run
-	var runId string
-
-	requestorParameter := ""
-	resultParameter := ""
-	fromAgeHours := 0
-	toAgeHours := 0
-	shouldGetActive := true
-
-	runs, err = GetRunsFromRestApi(runName, requestorParameter, resultParameter, fromAgeHours, toAgeHours, shouldGetActive, timeService, apiClient)
-
-	if err == nil {
-
-		if len(runs) > 1 {
-
-			// More than 1 active run has been found with this runName, as runs might be stuck in active state like ending...
-			// So find the run with the first startTime, and attempt to reset that one
-
-		} else if len(runs) == 1 {
-
-			runId = runs[0].GetRunId()
-
-		} else {
-
-			log.Printf("No active runs found matching run name: '%s'", runName)
-			err = galasaErrors.NewGalasaError(galasaErrors.GALASA_ERROR_NO_ACTIVE_RUNS_WITH_RUNNAME, runName)
-
-		}
-	}
-
-	return runId, err
-}
-
-func createUpdateRunStatusRequest(status string, result string) *galasaapi.UpdateRunStatusRequest {
-	var updateRunStatusRequest = galasaapi.NewUpdateRunStatusRequest()
-
-	updateRunStatusRequest.SetStatus(status)
-	updateRunStatusRequest.SetResult(result)
-
-	return updateRunStatusRequest
-}
